requirements.txt
streamlit
google-cloud-bigquery
google-cloud-storage
tensorflow
pandas
numpy
matplotlib
